refactor: organize directories
