+++ __tests__/main/index.test.ts
@@ -2,7 +2,7 @@
  * Unit tests for the action's entrypoint, src/index.ts
  */
 
-import * as main from '../src/main'
+import * as main from '../../src/main/main'
 
 // Mock the action's entrypoint
 const runMock = jest.spyOn(main, 'run').mockImplementation()
@@ -10,7 +10,7 @@ const runMock = jest.spyOn(main, 'run').mockImplementation()
 describe('index', () => {
   it('calls run when imported', async () => {
     // eslint-disable-next-line @typescript-eslint/no-require-imports
-    require('../src/index')
+    require('../../src/main/index')
 
     expect(runMock).toHaveBeenCalled()
   })
