+++ __tests__/main/input.test.ts
@@ -1,7 +1,7 @@
 import * as core from '@actions/core'
 import * as github from '@actions/github'
 
-import * as inputFunc from '../src/input'
+import * as inputFunc from '../../src/main/input'
 
 let getInputMock: jest.SpiedFunction<typeof core.getInput>
