+++ src/main/config.generated.ts
